+++ src/Engine/Rendering/Texture.cpp
@@ -4,15 +4,15 @@
 // Date: 10/11/2024
 //
 
-#include "Engine2D/Rendering/Texture2D.hpp"
+#include "Engine/Rendering/Texture.hpp"
 #include "Engine/RenderingHeaders.hpp"
 
-namespace Engine2D::Rendering {
-  Texture2D::Texture2D()
+namespace Engine::Rendering {
+  Texture::Texture()
     : id(0), width(0), height(0), internalFormat(GL_RGB), imageFormat(GL_RGB), wrapS(GL_CLAMP_TO_EDGE),
       wrapT(GL_CLAMP_TO_EDGE), filterMin(GL_NEAREST), filterMax(GL_NEAREST) {}
 
-  void Texture2D::generate(const int width, const int height, const unsigned char *data) {
+  void Texture::generate(const int width, const int height, const unsigned char *data) {
     this->width = width;
     this->height = height;
 
@@ -31,19 +31,19 @@ namespace Engine2D::Rendering {
     glBindTexture(GL_TEXTURE_2D, 0);
   }
 
-  void Texture2D::clear() {
+  void Texture::clear() {
     if (this->id > 0) {
       glDeleteTextures(1, &this->id);
       id = 0;
     }
   }
 
-  void Texture2D::bind() const {
+  void Texture::bind() const {
     glActiveTexture(GL_TEXTURE0);
     glBindTexture(GL_TEXTURE_2D, this->id);
   }
 
-  void Texture2D::unbind() {
+  void Texture::unbind() {
     glBindTexture(GL_TEXTURE_2D, 0);
   }
 }