+++ src/Engine2D/Entity2D.cpp
@@ -11,16 +11,15 @@
 #include "Engine2D/ParticleSystem/ParticleSystemRenderer2D.hpp"
 #include "Engine2D/ParticleSystem/ParticleSystem2D.hpp"
 #include "Engine2D/Physics/Physics2D.hpp"
-
-using Engine2D::Rendering::Texture2D;
+#include "Engine2D/Rendering/Renderer2D.hpp"
+#include "Engine2D/Rendering/SpriteRenderer.hpp"
 
 namespace Engine2D {
   Entity2D::Entity2D(std::string name)
-    : name(std::move(name)), transform(std::make_shared<Transform2D>()), textureColor(glm::vec4(1.0f)), active(true),
-      parentsActive(true), texture(nullptr) {}
+    : name(std::move(name)), transform(std::make_shared<Transform2D>()), active(true), parentsActive(true) {}
 
   bool Entity2D::operator==(const Entity2D &entity) const {
-    return name == entity.name && transform == entity.transform && texture == entity.texture;
+    return name == entity.name && transform == entity.transform;
   }
 
   bool Entity2D::operator!=(const Entity2D &entity) const {
@@ -35,23 +34,6 @@ namespace Engine2D {
 
   bool Entity2D::IsActive() const {
     return this->active && this->parentsActive;
-  }
-
-  void Entity2D::SetTexture(const std::shared_ptr<Texture2D> &texture) {
-    if (!texture || this->texture == texture)
-      return;
-
-    if (this->texture)
-      for (const auto it = Game2D::instance->entitiesToRender[this->texture->id].begin();
-           it != Game2D::instance->entitiesToRender[this->texture->id].end();)
-        if (*it == shared_from_this())
-          Game2D::instance->entitiesToRender[this->texture->id].erase(it);
-    this->texture = texture;
-    Game2D::instance->entitiesToRender[this->texture->id].emplace_back(shared_from_this());
-  }
-
-  std::shared_ptr<Texture2D> Entity2D::Texture() const {
-    return texture;
   }
 
   void Entity2D::Destroy() {
@@ -82,6 +64,8 @@ namespace Engine2D {
       Game2D::instance->physics2D->addCollider(collider);
     if (const auto particleSystem = std::dynamic_pointer_cast<ParticleSystem2D>(component))
       ParticleSystemRenderer2D::addParticleSystem(particleSystem);
+    if (const auto spriteRenderer = std::dynamic_pointer_cast<Rendering::SpriteRenderer>(component))
+      Rendering::Renderer2D::addRenderer(spriteRenderer);
   }
 
   void Entity2D::recallComponent(const std::shared_ptr<Component2D> &component) {
@@ -89,5 +73,7 @@ namespace Engine2D {
       Game2D::instance->physics2D->removeCollider(collider);
     if (const auto particleSystem = std::dynamic_pointer_cast<ParticleSystem2D>(component))
       ParticleSystemRenderer2D::removeParticleSystem(particleSystem);
+    if (const auto spriteRenderer = std::dynamic_pointer_cast<Rendering::SpriteRenderer>(component))
+      Rendering::Renderer2D::removeRenderer(spriteRenderer);
   }
 }