+++ Changelog.md
@@ -3,9 +3,20 @@
 All notable changes to this project will be documented in this file.<br>
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.5.4] - 2025-09-04
+### Added
+- **[Sprite]**: wrapper around Texture for rendering
+- **[SpriteRenderer]**: component used to render entity sprites
+### Changed
+- **[Texture2D--->Texture]**: now a simple class to load a texture
+- **[SpriteRenderer--->Renderer2D]**
+### Removed
+- **[Entity2D]**: no longer has texture and textureColor as member variables
+- **[Game2D]**: entitiesToRender are now stored in the Renderer2D class
+
 ## [0.5.3] - 2025-03-04
 ### Added
-- **[engine-cli]**: rebuild command rebuilds the cli's binary executable 
+- **[engine-cli]**: rebuild command rebuilds the cli's binary executable
 ### Changed
 - **[Transform2D]**: Set and Update functions all call the main one
 - **[engine-cli]**: commands must start with a "--" symbol now and multiple commands can be run together